Modularized and cleaned some code
