+++ CMakeLists.txt
@@ -1,5 +1,7 @@
 cmake_minimum_required(VERSION 3.2)
 
+include_directories( ${CMAKE_SOURCE_DIR}/include )
+
 set_source_files_properties("*.c" PROPERTIES LANGUAGE CXX )
 set_source_files_properties("*.h" PROPERTIES LANGUAGE CXX )
